+++ convex/convex.config.ts
@@ -3,9 +3,17 @@ import staticHosting from "@convex-dev/static-hosting/convex.config";
 import auth from "@estifanos-sh/convex-auth/convex.config";
 import { authEnv } from "@estifanos-sh/convex-auth/server";
 import { defineApp } from "convex/server";
+import { v } from "convex/values";
 
 const app = defineApp({
-  env: authEnv,
+  env: {
+    ...authEnv,
+    GOOGLE_CLIENT_ID: v.optional(v.string()),
+    GOOGLE_CLIENT_SECRET: v.optional(v.string()),
+    RESEND_API_KEY: v.string(),
+    SESSION_INACTIVE_DURATION_MS: v.optional(v.string()),
+    SESSION_TOTAL_DURATION_MS: v.optional(v.string()),
+  },
 });
 
 app.use(auth);